refactor: collect bits hints
