+++ std/hints.go
@@ -31,10 +31,7 @@ func registerHints() {
 	solver.RegisterHint(sw_bls12377.DecomposeScalarG1)
 	solver.RegisterHint(sw_bls24315.DecomposeScalarG2)
 	solver.RegisterHint(sw_bls12377.DecomposeScalarG2)
-	solver.RegisterHint(bits.NTrits)
-	solver.RegisterHint(bits.NNAF)
-	solver.RegisterHint(bits.IthBit)
-	solver.RegisterHint(bits.NBits)
+	solver.RegisterHint(bits.GetHints()...)
 	solver.RegisterHint(selector.GetHints()...)
 	solver.RegisterHint(emulated.GetHints()...)
 	solver.RegisterHint(rangecheck.GetHints()...)
