+++ std/math/bits/conversion_binary.go
@@ -3,15 +3,8 @@ package bits
 import (
 	"math/big"
 
-	"github.com/consensys/gnark/constraint/solver"
 	"github.com/consensys/gnark/frontend"
 )
-
-func init() {
-	// register hints
-	solver.RegisterHint(IthBit)
-	solver.RegisterHint(NBits)
-}
 
 // ToBinary is an alias of ToBase(api, Binary, v, opts)
 func ToBinary(api frontend.API, v frontend.Variable, opts ...BaseConversionOption) []frontend.Variable {
@@ -84,28 +77,4 @@ func toBinary(api frontend.API, v frontend.Variable, opts ...BaseConversionOptio
 	api.AssertIsEqual(Σbi, v)
 
 	return bits
-}
-
-// IthBit returns the i-tb bit the input. The function expects exactly two
-// integer inputs i and n, takes the little-endian bit representation of n and
-// returns its i-th bit.
-func IthBit(_ *big.Int, inputs []*big.Int, results []*big.Int) error {
-	result := results[0]
-	if !inputs[1].IsUint64() {
-		result.SetUint64(0)
-		return nil
-	}
-
-	result.SetUint64(uint64(inputs[0].Bit(int(inputs[1].Uint64()))))
-	return nil
-}
-
-// NBits returns the first bits of the input. The number of returned bits is
-// defined by the length of the results slice.
-func NBits(_ *big.Int, inputs []*big.Int, results []*big.Int) error {
-	n := inputs[0]
-	for i := 0; i < len(results); i++ {
-		results[i].SetUint64(uint64(n.Bit(i)))
-	}
-	return nil
 }
