+++ std/math/bits/conversion_ternary.go
@@ -4,17 +4,8 @@ import (
 	"math"
 	"math/big"
 
-	"github.com/consensys/gnark/constraint/solver"
 	"github.com/consensys/gnark/frontend"
 )
-
-// NTrits returns the first trits of the input. The number of returned trits is
-// defined by the length of the results slice.
-var NTrits = nTrits
-
-func init() {
-	solver.RegisterHint(NTrits)
-}
 
 // ToTernary is an alias of ToBase(api, Ternary, v, opts...)
 func ToTernary(api frontend.API, v frontend.Variable, opts ...BaseConversionOption) []frontend.Variable {
@@ -104,20 +95,4 @@ func AssertIsTrit(api frontend.API, v frontend.Variable) {
 	// TODO this adds 3 constraint, not 2. Need api.Compiler().AddConstraint(...)
 	y := api.Mul(api.Sub(1, v), api.Sub(2, v))
 	api.AssertIsEqual(api.Mul(v, y), 0)
-}
-
-func nTrits(_ *big.Int, inputs []*big.Int, results []*big.Int) error {
-	n := inputs[0]
-	// TODO using big.Int Text method is likely not cheap
-	base3 := n.Text(3)
-	i := 0
-	for j := len(base3) - 1; j >= 0 && i < len(results); j-- {
-		results[i].SetUint64(uint64(base3[j] - 48))
-		i++
-	}
-	for ; i < len(results); i++ {
-		results[i].SetUint64(0)
-	}
-
-	return nil
 }
